+++ Runtime/Ponito/Core/Tests/AsyncsTests.cs
@@ -5,6 +5,8 @@ using NUnit.Framework;
 using Ponito.Core.Asyncs.Extensions;
 // using Ponito.Core.Asyncs.Promises;
 using Ponito.Core.Asyncs.Tasks;
+using Ponito.Core.Extensions;
+using UnityEditor;
 using UnityEngine;
 using UnityEngine.TestTools;
 
@@ -15,12 +17,12 @@ namespace Ponito.Core.Tests
         [UnityTest]
         public IEnumerator TestMono()
         {
-            var go = new GameObject("AsyncsTestsMono");
-            var mono = go.AddComponent<AsyncsTestsMono>();
+            Selection.activeGameObject = new GameObject("AsyncsTestsMono").EnsureComponent(out AsyncsTestsMono mono);
             yield return mono.IsAllPassed();
             Assert.IsTrue(mono.isFinished);
+            yield return new WaitForSeconds(3f);
         }
-        
+
         [UnityTest]
         public IEnumerator TestDelay3000()
         {
@@ -34,37 +36,7 @@ namespace Ponito.Core.Tests
         {
             var before = Time.frameCount;
             yield return PoTask.Yield().WaitAsCoroutine();
-            var frameCount = Time.frameCount;
-            var frame      = frameCount - before;
-            Assert.IsTrue(frame == 1, $"frameCount - before = {frameCount} - {before} = {frame}");
-        }
-
-        [UnityTest]
-        public IEnumerator TestPoTaskDelay3000()
-        {
-            var before = Time.time;
-            yield return PoTaskDelay3000().WaitAsCoroutine();
-            Assert.IsTrue(Time.time - before > 3f);
-        }
-
-        [UnityTest]
-        public IEnumerator TestPoTaskYield50Frame()
-        {
-            var before = Time.frameCount;
-            yield return PoTaskYield50Frame().WaitAsCoroutine();
-            var frameCount = Time.frameCount;
-            var frame      = frameCount - before;
-            Assert.IsTrue(frame == 50, $"frameCount - before = {frameCount} - {before} = {frame}");
-        }
-
-        [UnityTest]
-        public IEnumerator TestPoTaskYield100Frame()
-        {
-            var before = Time.frameCount;
-            yield return PoTaskYield100Frame().WaitAsCoroutine();
-            var frameCount = Time.frameCount;
-            var frame      = frameCount - before;
-            Assert.IsTrue(frame == 100, $"frameCount - before = {frameCount} - {before} = {frame}");
+            Assert.That(Time.frameCount - before, Is.EqualTo(1));
         }
 
         [UnityTest]
@@ -87,52 +59,6 @@ namespace Ponito.Core.Tests
             Assert.NotNull(ex);
         }
 
-        // [UnityTest]
-        // public IEnumerator TestPoTask5()
-        // {
-        //     using var i = new Promise<int>();
-        //     yield return PoTask5().RunAsCoroutine(i);
-        //     Assert.IsTrue(i == 5);
-        // }
-        //
-        // [UnityTest]
-        // public IEnumerator TestPoTask10()
-        // {
-        //     using var i = new Promise<int>();
-        //     yield return PoTask10().RunAsCoroutine(i);
-        //     Assert.IsTrue(i == 10);
-        // }
-
-        [UnityTest]
-        public IEnumerator TestPoTaskCancel()
-        {
-            var before = Time.frameCount;
-
-            yield return PoTaskCancel()
-               .Try()
-               .WaitAsCoroutine();
-
-            var frameCount = Time.frameCount;
-            var frame      = frameCount - before;
-            Assert.IsTrue(frame == 5, $"frameCount - before = {frameCount} - {before} = {frame}");
-        }
-
-        private async PoTask PoTaskDelay3000()
-        {
-            await PoTask.Delay(3000);
-        }
-
-        private async PoTask PoTaskYield50Frame()
-        {
-            for (int i = 0; i < 50; i++) await PoTask.Yield();
-        }
-
-        private async PoTask PoTaskYield100Frame()
-        {
-            await PoTaskYield50Frame();
-            await PoTaskYield50Frame();
-        }
-
         private async PoTask PoTaskYieldException()
         {
             await PoTask.Yield();
@@ -142,25 +68,6 @@ namespace Ponito.Core.Tests
         private async PoTask PoTaskInnerException()
         {
             throw new Exception();
-        }
-
-        private async PoTask<int> PoTask5()
-        {
-            await PoTask.Delay(1f);
-            return 5;
-        }
-
-        private async PoTask<int> PoTask10()
-        {
-            return await PoTask5() + await PoTask5();
-        }
-
-        private async PoTask PoTaskCancel()
-        {
-            var cts = new CancellationTokenSource();
-            for (int i = 0; i < 5; i++) await PoTask.Yield();
-            cts.Cancel();
-            for (int i = 0; i < 5; i++) await PoTask.Yield(cts.Token);
         }
     }
 }